Co-locate concrete TypeName subtypes in their own files
We can do this now with sealed classes!
